Moved TimestampSigner and Serializer tests to their own files and retired Exercise test case.
