+++ phpunit.xml
@@ -1,4 +1,4 @@
-<phpunit colors="true">
+<phpunit bootstrap="itsdangerous.php" colors="true">
     <testsuites>
         <testsuite name="All">
             <directory>test/</directory>
